+++ src/features/userInfo/@hooks/useModal.ts
@@ -1,0 +1,15 @@
+import { useState } from 'react';
+
+export const useModal = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleOpenModal = () => {
+    setIsOpen(true);
+  };
+
+  const handleCloseModal = () => {
+    setIsOpen(false);
+  };
+
+  return { isOpen, handleOpenModal, handleCloseModal };
+};
